requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.11
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
onnx==1.17.0
opencv-contrib-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
pydantic_core==2.27.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
